rename files to line up with new names
