+++ src/levanter/data/__init__.py
@@ -1,4 +1,5 @@
 from levanter.data.dataset import Dataset, ShardableDataset, ShuffleDataset
+from levanter.data.loader import BatchLoader, ReplicatedBatchLoader, ShardedBatchLoader
 from levanter.data.utils import batched
 
 
@@ -7,4 +8,7 @@ __all__ = [
     "Dataset",
     "ShardableDataset",
     "ShuffleDataset",
+    "BatchLoader",
+    "ReplicatedBatchLoader",
+    "ShardedBatchLoader",
 ]
