+++ tests/test_replicated_loader.py
